Split applications & members from auth/data/groups/mod.rs
